refactor: seperate session model
